requirements.txt
fastAPI==0.0.7
ollama==0.8.0
openpyxl==3.1.5
python-docx==1.1.2
httpx==0.27.2
uvicorn==0.34.2